# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: taboo · 禁制 — 密钥管理平台 API
  version: 0.7.0
  description: 开源自部署密钥管理平台。高价值数据 + 严格访问控制 + 全程留痕。M5：Secret Sync / Webhooks / OIDC SSO / 审计导出。
servers:
  - url: http://localhost:7100
    description: 本地开发
...
paths:
...
  /api/v1/projects/{pid}/secrets:
    get:
      tags: [secrets]
      summary: 密钥列表（无值；?path= 按文件夹过滤，缺省全环境）
      operationId: listSecrets
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: path, in: query, schema: { type: string, description: 物化路径 /a/b/，缺省列出整个环境 } }
      responses:
        '200': { description: 密钥元数据列表, content: { application/json: { schema: { type: object, properties: { secrets: { type: array, items: { $ref: '#/components/schemas/SecretMeta' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [secrets]
      summary: 创建/更新密钥（AES-256-GCM 信封加密，产生新版本；同 key 不同目录互不冲突）
      operationId: upsertSecret
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: path, in: query, schema: { type: string, default: / } }
        - { name: Idempotency-Key, in: header, schema: { type: string }, description: 幂等约定（v1.5 强制） }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecretUpsertRequest' }
      responses:
        '200': { description: 已保存（含新版本号）, content: { application/json: { schema: { type: object, properties: { version: { type: integer } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/projects/{pid}/secrets/{key}:
    get:
      tags: [secrets]
      summary: 读取明文（reveal 权限，记审计；?path= 定位目录）
      operationId: revealSecret
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: key, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string, default: / } }
      responses:
        '200': { description: 明文, content: { application/json: { schema: { $ref: '#/components/schemas/SecretValue' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{pid}/secrets/{key}/versions:
    get:
      tags: [secrets]
      summary: 版本历史（只追加）
      operationId: secretVersions
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: key, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string, default: / } }
      responses:
        '200':
          description: 版本列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions: { type: array, items: { $ref: '#/components/schemas/SecretVersion' } }
                  latest: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{pid}/secrets/{key}/rollback:
    post:
      tags: [secrets]
      summary: 回滚到历史版本（以旧值产生新版本 v+1）
      operationId: rollbackSecret
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Pid' }
        - { $ref: '#/components/parameters/Env' }
        - { name: key, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string, default: / } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties: { version: { type: integer } }
      responses:
        '200': { description: 已回滚, content: { application/json: { schema: { type: object, properties: { version: { type: integer } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Pid: { name: pid, in: path, required: true, schema: { type: string } }
    Env: { name: env, in: query, schema: { type: string, default: dev } }
  responses:
    BadRequest: { description: 参数错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: 未认证/凭证错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: 权限不足（RBAC），如 viewer 取明文、developer 写 prod, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: 资源不存在, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: BAD_CREDENTIALS }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    SecretMeta:
      type: object
      properties:
        id: { type: string }
        folder: { type: string }
        key: { type: string }
        comment: { type: string }
        tags: { type: array, items: { type: string } }
        version: { type: integer }
        updated_at: { type: string }
        canReveal: { type: boolean }
        canWrite: { type: boolean }
    SecretValue:
      allOf:
        - { $ref: '#/components/schemas/SecretMeta' }
        - type: object
          properties: { value: { type: string } }
    SecretVersion:
      type: object
      properties:
        version: { type: integer }
        created_by: { type: string }
        created_at: { type: string }
    SecretUpsertRequest:
      type: object
      required: [key, value]
      properties:
        key: { type: string, pattern: '^[^/\s][^/]*[^/\s]$|^[^/\s]$', description: '单段 key，禁止包含 /（folder 与 key 分离，#16）' }
        value: { type: string }
        comment: { type: string }
        tags: { type: array, items: { type: string } }
...